EU-115 refactor educational_institutions
